DepositoBancario_01

+++ src/eser_cicli/DepositoBancario_01.java
@@ -6,11 +6,11 @@ public class DepositoBancario_01 {
         Scanner inp=new Scanner(System.in);
         double saldo_iniz,perc_interess,soglia,anni=0;
         System.out.println("inserire saldo iniziale");
-        saldo_iniz=inp.nextInt();
+        saldo_iniz=inp.nextDouble();
         System.out.println("inserire percentuale d'interesse (solo numero)");
-        perc_interess=inp.nextInt();
+        perc_interess=inp.nextDouble();
         System.out.println("inserire una soglia da raggiungere");
-        soglia=inp.nextInt();
+        soglia=inp.nextDouble();
         while(saldo_iniz<soglia){
             double percent=(saldo_iniz/100)*perc_interess;
             saldo_iniz+=percent;
